less stutter
