+++ src/atlas_search.rs
@@ -14,10 +14,11 @@ use crate::bson::{doc, Bson, Document};
 ///
 /// ```no_run
 /// # async fn wrapper() -> mongodb::error::Result<()> {
-/// # use mongodb::{Collection, atlas_search::AtlasSearch, bson::{Document, doc}};
+/// # use mongodb::{Collection, bson::{Document, doc}};
 /// # let collection: Collection<Document> = todo!();
+/// use mongodb::atlas_search;
 /// let cursor = collection.aggregate(vec![
-///     AtlasSearch::autocomplete("title", "pre")
+///     atlas_search::autocomplete("title", "pre")
 ///         .fuzzy(doc! { "maxEdits": 1, "prefixLength": 1, "maxExpansions": 256 })
 ///         .into(),
 ///     doc! {
@@ -55,14 +56,15 @@ impl<T> AtlasSearch<T> {
     /// of different types in a single `Vec`:
     /// ```no_run
     /// # async fn wrapper() -> mongodb::error::Result<()> {
-    /// # use mongodb::{Collection, atlas_search::AtlasSearch, bson::{Document, doc}};
+    /// # use mongodb::{Collection, bson::{Document, doc}};
     /// # let collection: Collection<Document> = todo!();
+    /// use mongodb::atlas_search;
     /// let cursor = collection.aggregate(vec![
-    ///     AtlasSearch::compound()
+    ///     atlas_search::compound()
     ///         .must(vec![
-    ///             AtlasSearch::text("description", "varieties").unit(),
-    ///             AtlasSearch::compound()
-    ///                 .should(AtlasSearch::text("description", "Fuji"))
+    ///             atlas_search::text("description", "varieties").unit(),
+    ///             atlas_search::compound()
+    ///                 .should(atlas_search::text("description", "Fuji"))
     ///                 .unit(),
     ///         ])
     ///         .into(),
@@ -280,166 +282,4 @@ impl AtlasSearch<Facet> {
         self.meta = false;
         self
     }
-}
-
-#[test]
-fn api_flow() {
-    assert_eq!(
-        doc! {
-            "$search": {
-                "autocomplete": {
-                    "query": "pre",
-                    "path": "title",
-                    "fuzzy": {
-                        "maxEdits": 1,
-                        "prefixLength": 1,
-                        "maxExpansions": 256,
-                    },
-                }
-            }
-        },
-        AtlasSearch::autocomplete("title", "pre")
-            .fuzzy(doc! { "maxEdits": 1, "prefixLength": 1, "maxExpansions": 256 })
-            .into()
-    );
-    assert_eq!(
-        doc! {
-            "$search": {
-                "text": {
-                    "path": "plot",
-                    "query": "baseball",
-                }
-            }
-        },
-        AtlasSearch::text("plot", "baseball").into()
-    );
-    assert_eq!(
-        doc! {
-            "$search": {
-                "compound": {
-                    "must": [{
-                        "text": {
-                            "path": "description",
-                            "query": "varieties",
-                        }
-                    }],
-                    "should": [{
-                        "text": {
-                            "path": "description",
-                            "query": "Fuji",
-                        }
-                    }],
-                }
-            }
-        },
-        AtlasSearch::compound()
-            .must(AtlasSearch::text("description", "varieties"))
-            .should(AtlasSearch::text("description", "Fuji"))
-            .into()
-    );
-    {
-        use short::*;
-        assert_eq!(
-            doc! {
-                "$search": {
-                    "embeddedDocument": {
-                        "path": "items",
-                        "operator": {
-                            "compound": {
-                                "must": [{
-                                    "text": {
-                                        "path": "items.tags",
-                                        "query": "school",
-                                    }
-                                }],
-                                "should": [{
-                                    "text": {
-                                        "path": "items.name",
-                                        "query": "backpack",
-                                    }
-                                }]
-                            }
-                        },
-                        "score": {
-                            "embedded": {
-                                "aggregate": "mean"
-                            }
-                        },
-                    }
-                }
-            },
-            embedded_document(
-                "items",
-                compound()
-                    .must(text("items.tags", "school"))
-                    .should(text("items.name", "backpack")),
-            )
-            .score(doc! {
-                "embedded": {
-                    "aggregate": "mean"
-                }
-            })
-            .into()
-        );
-    }
-    assert_eq!(
-        doc! {
-            "$search": {
-                "equals": {
-                    "path": "verified_user",
-                    "value": true,
-                }
-            }
-        },
-        AtlasSearch::equals("verified_user", true).into()
-    );
-    let gte_dt = crate::bson::DateTime::parse_rfc3339_str("2000-01-01T00:00:00.000Z").unwrap();
-    let lte_dt = crate::bson::DateTime::parse_rfc3339_str("2015-01-31T00:00:00.000Z").unwrap();
-    assert_eq!(
-        doc! {
-            "$searchMeta": {
-                "facet": {
-                    "operator": {
-                        "range": {
-                            "path": "released",
-                            "gte": gte_dt,
-                            "lte": lte_dt,
-                        }
-                    },
-                    "facets": {
-                        "directorsFacet": {
-                            "type": "string",
-                            "path": "directors",
-                            "numBuckets": 7,
-                        },
-                        "yearFacet": {
-                            "type": "number",
-                            "path": "year",
-                            "boundaries": [2000, 2005, 2010, 2015]
-                        },
-                    }
-                }
-            }
-        },
-        AtlasSearch::facet(doc! {
-            "directorsFacet": {
-                "type": "string",
-                "path": "directors",
-                "numBuckets": 7,
-            },
-            "yearFacet": {
-                "type": "number",
-                "path": "year",
-                "boundaries": [2000, 2005, 2010, 2015]
-            },
-        })
-        .operator(doc! {
-            "range": {
-                "path": "released",
-                "gte": gte_dt,
-                "lte": lte_dt,
-            }
-        })
-        .into()
-    );
 }
